# Vigen-Repository/ClientApp/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Vigen-Repository",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://localhost:44385"
    }
  ],
  "paths": {
    "/api/Notify": {
      "get": {
        "tags": [
          "Notify"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Notify"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Notify"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Notify"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Notify"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Notify"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Notify"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Notify"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Notify"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notify"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notify"
                }
              }
            }
          }
        }
      }
    },
    "/api/Notify/{id}": {
      "get": {
        "tags": [
          "Notify"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Notify"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notify"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notify"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Notify"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Notify"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Notify"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Notify"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Notify"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notify"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notify"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Notify"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Notify"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notify"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notify"
                }
              }
            }
          }
        }
      }
    },
    "/api/Organization": {
      "get": {
        "tags": [
          "Organization"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Organization"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Organization"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Organization"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Organization"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Organization"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Organization"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Organization"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        }
      }
    },
    "/api/Organization/{id}": {
      "get": {
        "tags": [
          "Organization"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Organization"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Organization"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Organization"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Organization"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Organization"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Notify": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "stateId": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "organizationTypeId": {
            "type": "string",
            "nullable": true
          },
          "organizationType": {
            "$ref": "#/components/schemas/OrganizationType"
          },
          "state": {
            "$ref": "#/components/schemas/State"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "additionalProperties": false
      },
      "Organization": {
        "type": "object",
        "properties": {
          "nit": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "tel": {
            "type": "string",
            "nullable": true
          },
          "sites": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Site"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OrganizationType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "notifies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notify"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Site": {
        "type": "object",
        "properties": {
          "nit": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "ubication": {
            "type": "string",
            "nullable": true
          },
          "range": {
            "type": "integer",
            "format": "int32"
          },
          "countryCode": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "tel": {
            "type": "string",
            "nullable": true
          },
          "nitNavigation": {
            "$ref": "#/components/schemas/Organization"
          }
        },
        "additionalProperties": false
      },
      "State": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "notifies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notify"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "User": {
        "type": "object",
        "properties": {
          "identification": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "birthdate": {
            "type": "string",
            "format": "date-time"
          },
          "countryCode": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "occupation": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "maritalStatus": {
            "type": "string",
            "nullable": true
          },
          "ubication": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "verification": {
            "type": "boolean"
          },
          "notifies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notify"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}